# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: sethukumarj.76@gmail.com
    name: API Support
    url: sethukumarj.com
  description: This is an Events Radar project. You can visit the GitHub repository
    at https://github.com/SethukumarJ/Events_Radar_Developement
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go + Gin Radar API
  version: "1.0"
paths:
...
  /admin/approve-event:
    patch:
      operationId: approves event
      parameters:
      - description: 'Event Name : '
        in: query
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: approves the event for admin
      tags:
      - Admin
...
  /admin/list-events:
    get:
      operationId: list all upcoming events
      parameters:
      - description: 'Page number: '
        in: query
        name: page
        required: true
        type: integer
      - description: 'Page capacity : '
        in: query
        name: pagesize
        required: true
        type: integer
      - description: 'List event by approved non approved : '
        in: query
        name: approved
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: list all upcoming events for admin
      tags:
      - Admin
  /admin/list-organizations:
    get:
      operationId: list all organization with status
      parameters:
      - description: 'Page number: '
        in: query
        name: page
        required: true
        type: integer
      - description: 'Page capacity : '
        in: query
        name: pagesize
        required: true
        type: integer
      - description: 'List organization based on status: '
        in: query
        name: applicationStatus
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: list all pending organizations for admin
      tags:
      - Admin
  /admin/list-users:
    get:
      operationId: list all active users
      parameters:
      - description: 'Page number: '
        in: query
        name: page
        required: true
        type: string
      - description: 'Page capacity : '
        in: query
        name: pagesize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: list all active users for admin
      tags:
      - Admin
...
  /admin/make/vip-user:
    patch:
      operationId: make vip user
      parameters:
      - description: 'User Name : '
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: makes the user vip
      tags:
      - Admin
  /admin/register-organization:
    patch:
      operationId: Register organization
      parameters:
      - description: 'orgStatus id : '
        in: query
        name: orgstatusid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Resginter the organization
      tags:
      - Admin
  /admin/reject-organization:
    patch:
      operationId: Reject organization
      parameters:
      - description: 'orgStatus id : '
        in: query
        name: orgstatusid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Rejects the organization
      tags:
      - Admin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"